// DigitalHealth/src/app/models/ki-model.ts
export interface KI {
  referenceId?: String;
  _id: String;
  computer: Number;
  bewegung: Number;
  auto: Number;
  heben: Number;
  sitzen: Number;
  schmerzenRuecken: number;

  schmerzenNacken: number;

  schmerzenHandgelenk: number;

  schmerzenKnie: number;

  schmerzenHuefte: number;

  schmerzenSchulter: number;

  schmerzenFussgelenke: number;

  toggelBeweglichket: Boolean;

  schmerzenBrennen: Boolean;

  schmerzenKribbeln: Boolean;

  schmerzenTaubheit: Boolean;

  schmerzenUeberempfindlichkeit: Boolean;

  toggelkraft: Boolean;

  schmerzenNadeln: Boolean;

  schmerzenFeld: String;

  schmerzenArbeiten: Boolean;

  schmerzenLaufen: Boolean;

  schmerzenBuecken: Boolean;

  schmerzenStress: Boolean;

  schmerzenWetter: Boolean;

  schmerzenSpringen: Boolean;
  Beine: number;
  Ruecken: number;
  Oberkoeper: number;
}